+++ sdk/storage/azure-storage-files-datalake/inc/azure/storage/files/datalake/datalake_directory_client.hpp
@@ -160,24 +160,65 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     /**
      * @brief Sets POSIX access control rights on files and directories under given directory
      * recursively.
-     * @param mode Mode PathSetAccessControlRecursiveMode::Set sets POSIX access control rights on
-     * files and directories, PathSetAccessControlRecursiveMode::Modify modifies one or more POSIX
-     * access control rights  that pre-exist on files and directories,
-     * PathSetAccessControlRecursiveMode::Remove removes one or more POSIX access control rights
-     * that were present earlier on files and directories
      * @param acls Sets POSIX access control rights on files and directories. Each access control
      * entry (ACE) consists of a scope, a type, a user or group identifier, and permissions.
      * @param options Optional parameters to set an access control recursively to the resource the
      * directory points to.
-     * @return Azure::Core::Response<Models::SetDataLakeDirectoryAccessControlRecursiveResult>
+     * @return
+     * Azure::Core::Response<Models::SetDataLakeDirectoryAccessControlRecursiveSinglePageResult>
      * @remark This request is sent to dfs endpoint.
      */
-    Azure::Core::Response<Models::SetDataLakeDirectoryAccessControlRecursiveResult>
-    SetAccessControlRecursive(
-        Models::PathSetAccessControlRecursiveMode mode,
+    Azure::Core::Response<Models::SetDataLakeDirectoryAccessControlRecursiveSinglePageResult>
+    SetAccessControlRecursiveSinglePage(
         std::vector<Models::Acl> acls,
-        const SetDataLakeDirectoryAccessControlRecursiveOptions& options
-        = SetDataLakeDirectoryAccessControlRecursiveOptions()) const;
+        const SetDataLakeDirectoryAccessControlRecursiveSinglePageOptions& options
+        = SetDataLakeDirectoryAccessControlRecursiveSinglePageOptions()) const
+    {
+      return SetAccessControlRecursiveSinglePageInternal(
+          Models::PathSetAccessControlRecursiveMode::Set, acls, options);
+    }
+
+    /**
+     * @brief Updates POSIX access control rights on files and directories under given directory
+     * recursively.
+     * @param acls Updates POSIX access control rights on files and directories. Each access control
+     * entry (ACE) consists of a scope, a type, a user or group identifier, and permissions.
+     * @param options Optional parameters to set an access control recursively to the resource the
+     * directory points to.
+     * @return
+     * Azure::Core::Response<Models::UpdateDataLakeDirectoryAccessControlRecursiveSinglePageResult>
+     * @remark This request is sent to dfs endpoint.
+     */
+    Azure::Core::Response<Models::UpdateDataLakeDirectoryAccessControlRecursiveSinglePageResult>
+    UpdateAccessControlRecursiveSinglePage(
+        std::vector<Models::Acl> acls,
+        const UpdateDataLakeDirectoryAccessControlRecursiveSinglePageOptions& options
+        = UpdateDataLakeDirectoryAccessControlRecursiveSinglePageOptions()) const
+    {
+      return SetAccessControlRecursiveSinglePageInternal(
+          Models::PathSetAccessControlRecursiveMode::Modify, acls, options);
+    }
+
+    /**
+     * @brief Removes POSIX access control rights on files and directories under given directory
+     * recursively.
+     * @param acls Removes POSIX access control rights on files and directories. Each access control
+     * entry (ACE) consists of a scope, a type, a user or group identifier, and permissions.
+     * @param options Optional parameters to set an access control recursively to the resource the
+     * directory points to.
+     * @return
+     * Azure::Core::Response<Models::RemoveDataLakeDirectoryAccessControlRecursiveSinglePageResult>
+     * @remark This request is sent to dfs endpoint.
+     */
+    Azure::Core::Response<Models::RemoveDataLakeDirectoryAccessControlRecursiveSinglePageResult>
+    RemoveAccessControlRecursiveSinglePage(
+        std::vector<Models::Acl> acls,
+        const RemoveDataLakeDirectoryAccessControlRecursiveSinglePageOptions& options
+        = RemoveDataLakeDirectoryAccessControlRecursiveSinglePageOptions()) const
+    {
+      return SetAccessControlRecursiveSinglePageInternal(
+          Models::PathSetAccessControlRecursiveMode::Remove, acls, options);
+    }
 
     /**
      * @brief List the paths in this file system.
@@ -200,6 +241,14 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
         : DataLakePathClient(std::move(dfsUrl), std::move(blobClient), pipeline)
     {
     }
+
+    Azure::Core::Response<Models::SetDataLakeDirectoryAccessControlRecursiveSinglePageResult>
+    SetAccessControlRecursiveSinglePageInternal(
+        Models::PathSetAccessControlRecursiveMode mode,
+        std::vector<Models::Acl> acls,
+        const SetDataLakeDirectoryAccessControlRecursiveSinglePageOptions& options
+        = SetDataLakeDirectoryAccessControlRecursiveSinglePageOptions()) const;
+
     friend class DataLakeFileSystemClient;
   };
 }}}} // namespace Azure::Storage::Files::DataLake