+++ sdk/storage/azure-storage-files-datalake/CHANGELOG.md
@@ -15,6 +15,8 @@
 - Added `Metadata`, `AccessType`, `HasImmutabilityPolicy`, `HasLegalHold`, `LeaseDuration`, `LeaseState` and `LeaseStatus` to `FileSystemItem`.
 - Added new type `LeaseDurationType` to indicate if a lease duration is fixed or infinite.
 - Added `RequestId` in each return type for REST API calls, except for concurrent APIs.
+- Added `UpdateAccessControlRecursiveSinglePage` to update the access control recursively for a datalake directory.
+- Added `RemoveAccessControlRecursiveSinglePage` to remove the access control recursively for a datalake directory.
 
 ### Breaking Changes
 
@@ -32,6 +34,8 @@
 - Changed all previous `LeaseDuration` members to a new type named `LeaseDurationType`.
 - `startsOn` parameter for `GetUserDelegationKey` was changed to optional.
 - Removed `PreviousContinuationToken` from `ListFileSystemsSinglePageResult`.
+- Changed `SetAccessControlRecursive` to `SetAccessControlRecursiveSinglePage`, to mark that it is a single page operation, and removed the `mode` parameter, separated the modify/delete functionality to two new APIs.
+- Changed `MaxRecord` to `MaxEntries`, `ForceFlag` to `ContinueOnFailure` to be more accurate names.
 
 ### Other Changes and Improvements